Removed Word component merging logic from controller to WordDownloader class
