+++ Epsilon.Abstractions/Component/IWordDownloader.cs
@@ -1,0 +1,6 @@
+﻿namespace Epsilon.Abstractions.Component;
+
+public interface IWordDownloader
+{
+    public Task<Stream> Download();
+}
