+++ Epsilon.Host.WebApi/Epsilon.Host.WebApi.csproj
@@ -14,4 +14,10 @@
       <ProjectReference Include="..\Epsilon\Epsilon.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Content Update="appsettings.Development.json">
+        <CopyToPublishDirectory>Never</CopyToPublishDirectory>
+      </Content>
+    </ItemGroup>
+
 </Project>
